refactor util functions
