+++ src/components/ACSClusterDefinitionForm.tsx
@@ -1,8 +1,8 @@
 import hljs from "highlight.js";
-import { Map, Seq } from "immutable";
 import { JSONSchema6 } from "json-schema";
 import React from "react";
 import Form from "react-jsonschema-form";
+import { Util } from "../shared/util";
 import { IClusterDefinition } from "../types";
 
 /**
@@ -756,48 +756,13 @@ const k8sSchema: JSONSchema6 = {
   type: "object",
 };
 
-/**
- * Return a list key lists for all object paths
- * @param {{[p: string]: any}} map
- * @param {string[]} currentPath
- * @returns {string[][]}
- */
-const getObjectKeyPaths = (map: { [key: string]: any }, currentPath: string[] = []): string[][] => {
-  const entries = Seq(Object.entries(map));
-  return [
-    ...entries.map(([k]) => [...currentPath, k]),
-    ...entries
-      .filter(([_, v]) => typeof v === "object")
-      .flatMap(([k]) => getObjectKeyPaths(map[k], [...currentPath, k])),
-  ].sort();
-};
-
-/**
- * Return a new object with all keys pointing to empty objects removed
- * @param {{[p: string]: any}} object
- * @returns {{[p: string]: any}}
- */
-const removeEmptyObjects = (object: { [key: string]: any }): { [key: string]: any } => {
-  const map = Map({ ...object });
-
-  return Seq(getObjectKeyPaths(object))
-    .map(keyPath => [keyPath, map.getIn(keyPath)])
-    .filter(
-      ([_, value]) => typeof value !== "object" || (typeof value === "object" && value.length > 0),
-    )
-    .reduce((carry, [keyPath, value]) => {
-      return carry.setIn(keyPath, value);
-    }, Map({}))
-    .toJS();
-};
-
 interface IACSClusterDefinitionFormJSON {
   clusterDefinition: IClusterDefinition;
 }
 
 export const ACSClusterDefinitionFormJSON = (props: IACSClusterDefinitionFormJSON) => {
   const { clusterDefinition } = props;
-  const cleaned = removeEmptyObjects(clusterDefinition);
+  const cleaned = Util.removeEmptyObjects(clusterDefinition);
 
   return (
     <div className="JSON">
